requirements.txt
alembic==1.7.7
anyio==3.6.1
APScheduler==3.9.1
cachetools==5.0.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
greenlet==1.1.2
h11==0.12.0
httpcore==0.14.7
httpx==0.22.0
idna==3.3
Mako==1.2.0
MarkupSafe==2.1.1
psycopg2==2.9.3
python-decouple==3.6
python-telegram-bot==20.0a0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.36
tornado==6.1
tzdata==2022.1
tzlocal==4.2
